rename double named services and pipes
