+++ YouTube-client-app/src/app/core/components/header/search-input/search-input.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
 import { ButtonComponent } from '../../../../shared/components/button/button.component';
-import { CardShowService } from '../../../../youtube/services/card-show-service.service';
+import { CardShowService } from '../../../../youtube/services/card-show.service';
 import { SearchStateService } from '../../../services/search-state.service';
 
 @Component({
