+++ YouTube-client-app/src/app/core/pipes/search.pipe.ts
