+++ YouTube-client-app/src/app/youtube/pipes/filter-views.pipe.ts
